cargo: Update GTK and related libraries
This requires quite a bit of churn, because many APIs have been
changed to be more ergonomic or to better expose the features of
the underlying toolkit.

One unfortunate side-effect is that we need to introduce a couple
of tiny unsafe sections: getting rid of them will require
reworking how we use dialogs, and that's something that will have
to wait for now.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>