src: Set application ID via gtk::Application
This ID is used to uniquely identify the application, so it's a
good idea to set it even though we don't use D-Bus.

At the same time, stop setting the application name via GLib,
which doesn't provide any additional benefit, and setting the
WM class for the window, which is something that requires using
deprecated GTK functionality and becomes irrelevant anyway if
GNOME can find the matching .desktop file.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>