build: Simplify build system
Flatpak is very neat, but the way we're using it at the moment
is kind of a mess and makes building the application fragile.

Use a minimalistic Makefile that contains a lot of assumptions
instead. Eventually we're going to switch to Meson, but for now
this will do the trick.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>