src: Improve search functionality
Currently the Ctrl+f shortcut toggles search on and off, same as
the search button: this is unlike most other applications, where
the button toggles search but using the keyboard shortcut always
activates search mode, and if pressed again while search mode is
already active it focuses the search field.

Change the search functionality so that it works the way users
expect based on what other applications do.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>